+++ src/processor/ast.ts
@@ -30,55 +30,55 @@ export function transformer(
                 // 1. Find callee function
                 if (path.get("callee").isIdentifier({ name })) {
                     path.traverse({
-                        CallExpression(path) {
-                            if (path.get("callee").isIdentifier({ name })) {
-                                path.replaceWith(
-                                    types.arrayExpression(path.node.arguments as types.Expression[])
+                        CallExpression(innerPath) {
+                            if (innerPath.get("callee").isIdentifier({ name })) {
+                                innerPath.replaceWith(
+                                    types.arrayExpression(innerPath.node.arguments as types.Expression[])
                                 )
                             }
                         }
                     })
 
                     path.traverse({
-                        ObjectExpression(path) {
+                        ObjectExpression(innerPath) {
                             try {
                                 // 2. Find outer Object(s) inside callee function
-                                if (!path.findParent((parentPath) => parentPath.isObjectExpression())) {
+                                if (!innerPath.findParent(parentPath => parentPath.isObjectExpression() && parentPath.node === innerPath.node)) {
                                     // 3. First parse ternary and logical conditionals
-                                    path.traverse({
-                                        ConditionalExpression(path) {
-                                            if (path.findParent((parent) => parent.isTemplateLiteral())) {
-                                                path.replaceWith([
-                                                    path.node.consequent, types.stringLiteral(" "), path.node.alternate
+                                    innerPath.traverse({
+                                        ConditionalExpression(innerPath) {
+                                            if (innerPath.findParent((parent) => parent.isTemplateLiteral())) {
+                                                innerPath.replaceWith([
+                                                    innerPath.node.consequent, types.stringLiteral(" "), innerPath.node.alternate
                                                 ].reduce((acc, current) => types.binaryExpression("+", acc, current)))
                                             } else {
-                                                path.replaceWith(
-                                                    types.arrayExpression([path.node.consequent, path.node.alternate])
+                                                innerPath.replaceWith(
+                                                    types.arrayExpression([innerPath.node.consequent, innerPath.node.alternate])
                                                 )
                                             }
                                         },
-                                        LogicalExpression(path) {
-                                            path.replaceWith(path.node.right)
+                                        LogicalExpression(innerPath) {
+                                            innerPath.replaceWith(innerPath.node.right)
                                         }
                                     })
 
                                     // 4. And parse the values if it is conditional
-                                    path.traverse({
-                                        ObjectProperty(path) {
+                                    innerPath.traverse({
+                                        ObjectProperty(innerPath) {
                                             if (
-                                                !path.node.shorthand
-                                                && !types.isStringLiteral(path.node.value)
-                                                && !types.isArrayExpression(path.node.value)
-                                                && !types.isTemplateLiteral(path.node.value)
-                                            ) path.node.value = types.numericLiteral(1)
+                                                !innerPath.node.shorthand
+                                                && !types.isStringLiteral(innerPath.node.value)
+                                                && !types.isArrayExpression(innerPath.node.value)
+                                                && !types.isTemplateLiteral(innerPath.node.value)
+                                            ) innerPath.node.value = types.numericLiteral(1)
                                         }
                                     })
 
                                     // 5. Then take the outer Object(s) out
-                                    const largestObject = generate(path.node).code
+                                    const largestObject = generate(innerPath.node).code
 
                                     // DONE. Final replace the original outer Object(s) with parsed one
-                                    path.replaceWith(types.stringLiteral(
+                                    innerPath.replaceWith(types.stringLiteral(
                                         parser(options)(...new Function(
                                             `return [${(/['"`]|:\s*1/g).test(largestObject) ? largestObject : ""}]`
                                         )() as ClassValue[])
@@ -97,3 +97,31 @@ export function transformer(
 
     return generate(ast).code
 }
+
+const text = `
+    const [toastButtons, setToastButtons] = useState<ToastButtonProps[]>([
+        {
+            idx: "1",
+            message: "Successful Toast",
+            type: toast.success,
+            styles: {
+                before: "abc",
+                after: cn(
+                    "col-start-1", {
+                    before: [
+                        "bg-mega-secondary opacity-10 transition-opacity",
+                        {
+                            dark: "bg-mega-success",
+                            hover: "opacity-20 duration-0"
+                        }
+                    ]
+                })
+            },
+            icon: <FaCircleCheck size={28} className="size-7 dark:fill-mega-success" />
+        }
+    ])
+`
+
+console.log(transformer(text, {
+    callee: "cn"
+}))